refactor(docs): colocate DocsArticle and DocsTocList styles
DocsShell.module.scss carried every docs style; per the SCSS colocation convention each
component's rules now live beside its .tsx. Rules moved unrenamed: article/markdown rendering,
callouts, code blocks, the markdown menu, TOC panes, and the article half of the print block go
to DocsArticle.module.scss; the TOC link styles to DocsTocList.module.scss. The shell keeps the
layout panes, sidebar, search, compact bar, and the chrome-hiding print rules. DocsArticle still
imports the shell module solely for the two shared chrome-button classes
(.sidebarIconBtn/.sidebarClose), which stay defined once there.
